Fix trivial compilation issues on BE machines
This eliminates some issues flagged by the compiler when compiling for a big-endian
machine:

* warnings in MD5 that would not ultimately cause problems;
* warnings in hand-crafted CDR that would result in incorrect data used in test cases;
* erors in security tests due to a typo in a byte-order selection constant.


Eliminate warnings on big-endian machines


Fix typo in endianness selection in big-endian machines

+++ src/core/cdr/test/mem_ser.h
@@ -25,14 +25,14 @@
   (unsigned char)( (uint32_t)(v)        & 0xff)
 #define SER32BE(v) SER32(v)
 #define SER64(v) \
-  (unsigned char)( (uint32_t)(v) >> 56),         \
-  (unsigned char)(((uint32_t)(v) >> 48) & 0xff), \
-  (unsigned char)(((uint32_t)(v) >> 40) & 0xff), \
-  (unsigned char)(((uint32_t)(v) >> 32) & 0xff), \
-  (unsigned char)(((uint32_t)(v) >> 24) & 0xff), \
-  (unsigned char)(((uint32_t)(v) >> 16) & 0xff), \
-  (unsigned char)(((uint32_t)(v) >>  8) & 0xff), \
-  (unsigned char)( (uint32_t)(v)        & 0xff)
+  (unsigned char)( (uint64_t)(v) >> 56),         \
+  (unsigned char)(((uint64_t)(v) >> 48) & 0xff), \
+  (unsigned char)(((uint64_t)(v) >> 40) & 0xff), \
+  (unsigned char)(((uint64_t)(v) >> 32) & 0xff), \
+  (unsigned char)(((uint64_t)(v) >> 24) & 0xff), \
+  (unsigned char)(((uint64_t)(v) >> 16) & 0xff), \
+  (unsigned char)(((uint64_t)(v) >>  8) & 0xff), \
+  (unsigned char)( (uint64_t)(v)        & 0xff)
 #define SER64BE(v) SER64(v)
 #else
 #define SER16(v) \
